chore: document types and methods
