+++ packages/core/src/device.ts
@@ -1,3 +1,7 @@
+/**
+ * The expected interface for a HIDDevice.
+ * This is to be implemented by any wrapping libraries to translate their platform specific devices into a common and simpler form
+ */
 export interface HIDDevice {
 	dataKeyOffset?: number
