+++ packages/core/src/models/base.ts
@@ -3,6 +3,7 @@ import { EventEmitter } from 'events'
 import { HIDDevice } from '../device'
 import { DeviceModelId } from '../models'
 import { KeyIndex } from './id'
+import { FillImageOptions, FillPanelOptions, StreamDeck } from './types'
 
 export type EncodeJPEGHelper = (buffer: Buffer, width: number, height: number) => Promise<Buffer>
 
@@ -20,103 +21,9 @@ export interface StreamDeckProperties {
 	KEY_DATA_OFFSET: number
 }
 
-export interface FillImageOptions {
-	format: 'rgb' | 'rgba' | 'bgr' | 'bgra'
-}
-export type FillPanelOptions = FillImageOptions
-
 export interface InternalFillImageOptions extends FillImageOptions {
 	offset: number
 	stride: number
-}
-
-export interface StreamDeck {
-	readonly NUM_KEYS: number
-	readonly KEY_COLUMNS: number
-	readonly KEY_ROWS: number
-
-	readonly ICON_SIZE: number
-	readonly ICON_PIXELS: number
-	readonly ICON_BYTES: number
-
-	readonly MODEL: DeviceModelId
-
-	/**
-	 * Checks if a keyIndex is valid. Throws an error on failure
-	 * @param keyIndex The key to check
-	 */
-	checkValidKeyIndex(keyIndex: KeyIndex): void
-
-	/**
-	 * Close the device
-	 */
-	close(): Promise<void>
-
-	/**
-	 * Fills the given key with a solid color.
-	 *
-	 * @param {number} keyIndex The key to fill
-	 * @param {number} r The color's red value. 0 - 255
-	 * @param {number} g The color's green value. 0 - 255
-	 * @param {number} b The color's blue value. 0 -255
-	 */
-	fillKeyColor(keyIndex: KeyIndex, r: number, g: number, b: number): Promise<void>
-
-	/**
-	 * Fills the given key with an image in a Buffer.
-	 *
-	 * @param {number} keyIndex The key to fill
-	 * @param {Buffer} imageBuffer
-	 * @param {Object} options
-	 */
-	fillKeyBuffer(keyIndex: KeyIndex, imageBuffer: Buffer, options?: FillImageOptions): Promise<void>
-
-	/**
-	 * Fills the whole panel with an image in a Buffer.
-	 *
-	 * @param {Buffer} imageBuffer
-	 */
-	fillPanelBuffer(imageBuffer: Buffer, options?: FillPanelOptions): Promise<void>
-
-	/**
-	 * Clears the given key.
-	 *
-	 * @param {number} keyIndex The key to clear
-	 */
-	clearKey(keyIndex: KeyIndex): Promise<void>
-
-	/**
-	 * Clears all keys.
-	 */
-	clearPanel(): Promise<void>
-
-	/**
-	 * Sets the brightness of the keys on the Stream Deck
-	 *
-	 * @param {number} percentage The percentage brightness
-	 */
-	setBrightness(percentage: number): Promise<void>
-
-	/**
-	 * Resets the display to the startup logo
-	 */
-	resetToLogo(): Promise<void>
-
-	/**
-	 * Get firmware version from Stream Deck
-	 */
-	getFirmwareVersion(): Promise<string>
-
-	/**
-	 * Get serial number from Stream Deck
-	 */
-	getSerialNumber(): Promise<string>
-
-	on(
-		...args:
-			| [event: 'down' | 'up', listener: (keyIndex: KeyIndex) => void]
-			| [event: 'error', listener: (e: unknown) => void]
-	): StreamDeck
 }
 
 export abstract class StreamDeckBase extends EventEmitter implements StreamDeck {
