+++ packages/node/src/device.ts
@@ -2,12 +2,22 @@ import { DeviceModelId, HIDDevice } from '@elgato-stream-deck/core'
 import { EventEmitter } from 'events'
 import * as HID from 'node-hid'
 
+/**
+ * Information about a found streamdeck
+ */
 export interface StreamDeckDeviceInfo {
+	/** The model of the device */
 	model: DeviceModelId
+	/** The connected path of the device in the usb tree */
 	path: string
+	/** The serialNumber of the device. If set it can be used as a unique hardware identifier */
 	serialNumber?: string
 }
 
+/**
+ * The wrapped node-hid HIDDevice.
+ * This translates it into the common format expected by @elgato-stream-deck/core
+ */
 export class NodeHIDDevice extends EventEmitter implements HIDDevice {
 	public dataKeyOffset?: number
 	private device: HID.HID
