+++ packages/webhid/src/index.ts
@@ -7,6 +7,10 @@ import { StreamDeckWeb } from './wrapper'
 export { DeviceModelId, KeyIndex, StreamDeck } from '@elgato-stream-deck/core'
 export { StreamDeckWeb } from './wrapper'
 
+/**
+ * Request the user to select some streamdecks to open
+ * @param userOptions Options to customise the device behvaiour
+ */
 export async function requestStreamDecks(options?: OpenStreamDeckOptions): Promise<StreamDeckWeb[]> {
 	// TODO - error handling
 	return navigator.hid
@@ -22,6 +26,23 @@ export async function requestStreamDecks(options?: OpenStreamDeckOptions): Promi
 		})
 }
 
+/**
+ * Reopen previously selected streamdecks.
+ * The browser remembers what the user previously allowed your site to access, and this will open those without the request dialog
+ * @param options Options to customise the device behvaiour
+ */
+export async function getStreamDecks(options?: OpenStreamDeckOptions): Promise<StreamDeckWeb[]> {
+	// TODO - error handling
+	return navigator.hid.getDevices().then(async (browserDevices) => {
+		return Promise.all(browserDevices.map((dev) => openDevice(dev, options)))
+	})
+}
+
+/**
+ * Open a StreamDeck from a manually selected HIDDevice handle
+ * @param browserDevice The unopened browser HIDDevice
+ * @param userOptions Options to customise the device behvaiour
+ */
 export async function openDevice(
 	browserDevice: HIDDevice,
 	userOptions?: OpenStreamDeckOptions
@@ -49,12 +70,4 @@ export async function openDevice(
 
 	const device: StreamDeck = new model.class(new WebHIDDevice(browserDevice), options || {})
 	return new StreamDeckWeb(device)
-}
-
-// getStreamDeck returns a streamdeck that was previously selected.
-export async function getStreamDecks(options?: OpenStreamDeckOptions): Promise<StreamDeckWeb[]> {
-	// TODO - error handling
-	return navigator.hid.getDevices().then(async (browserDevices) => {
-		return Promise.all(browserDevices.map((dev) => openDevice(dev, options)))
-	})
 }
